Add a fingerprint column with a unique constraint that contains dob + first_name + last_name
Unfortunately, postgres does not allow a case-insensitive multiple-column unique constraint,
nor does it allow "insert ... on conflict" to use an index name, only a constraint name.
So it seems the only way to get a case-insensitive multiple column upsert working is to
either create a new column and put your own value in there (here, a lowercased concatenation
of 3 fields), or add citext versions of the columns you care about and make a multi-column
unique constraint.