Build one image and deploy to all targets
Build the image as part of the test stage so that it happens in parallel
with the test runner. That way we don't have to wait for the tests to
finish running before building the image. Deploy still happens after
the test stage, and requires both test runner and image builder to pass.