show all phone numbers of a contact in case investigation view
a contact might have multiple phone numbers if they got edited
elsewhere. the contact form only shows one and edits are additive.
but you can now see all of the numbers when the edit is successful.
which is imperfect, but easier to understand and work around than
if the new phone number you added is not visible becaue only  the old
one shows.

[#175693747]