log all the contacts viewed in a case investigation
the profile for a person can have multiple case investigations
each of which can have multiple contacts. We show some personal
details for each of those contacts, so log that this user saw those
people.

Not to be confused with the "contact investigation" section which
can also sometimes appear independently on the profile page.

[#176078482]

Signed-off-by: Pat Maddox <pat@geometer.io>